nJoy: Fixed the analog trigger buttons
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1934 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/PluginPAD.h
@@ -6,7 +6,7 @@
 
 namespace Common {
     typedef void (__cdecl* TPAD_GetStatus)(u8, SPADStatus*);
-    typedef void (__cdecl* TPAD_Input)(u8, u8);
+    typedef void (__cdecl* TPAD_Input)(u16, u8);
     typedef void (__cdecl* TPAD_Rumble)(u8, unsigned int, unsigned int);
     typedef unsigned int (__cdecl* TPAD_GetAttachedPads)();
