+++ Source/Plugins/Plugin_nJoy_Testing/Src/nJoy.cpp
@@ -273,7 +273,7 @@ void DoState(unsigned char **ptr, int mode) {
 
 // Set buttons status from wxWidgets in the main application
 // ¯¯¯¯¯¯¯¯¯¯¯¯¯¯
-void PAD_Input(u8 _Key, u8 _UpDown) {}
+void PAD_Input(u16 _Key, u8 _UpDown) {}
